requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
protobuf==4.21.12
psycopg==3.1.12
PyJWT==2.8.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
